Requests: Fix patterns. (#968)
* Azure: Fix request patterns.

The old pattern was intended for versions 16.0.3, 16.1.3, 17.0.1 or 17.1.0,
but actually it only applies to 17.1.1. See the following snippet for example:

https://go.cluster.fun/snippet/SmLjQqzcUO_q

Hint: Better don't jump from kube-state-metrics 1.5.0 to 1.10.0 in 16.x releases.

* AWS: Fix request patterns.

The old pattern was intended for versions 16.0.2, 16.1.1 or 16.2.1,
but actually it only applies to 16.2.1. See the following snippet for example:

https://go.cluster.fun/snippet/lJsH75QB1bO

* AWS: Fix request patterns.

The old pattern was intended for versions 16.0.2, 16.1.1 or 16.3.2,
but actually it only applies to 16.3.2. See the following snippet for example:

https://go.cluster.fun/snippet/rpttRUUbZbV

* AWS: Fix request patterns.

The old pattern was intended for versions 16.1.2, 16.3.2, 16.4.2, 17.1.1 or 17.2.1,
but actually it only applies to 17.2.1. See the following snippet for example:

https://go.cluster.fun/snippet/vke0St2Z1VD

* Requests: Update cert-manager & kube-state-metrics as intended before.